Factor out SenderGangOfGangs into separate file
